Skip First Two Items in RemoveUnwantedItems Method
When removing html elements we don't want, keep the first two items of
the elements array regardless if they are a bad item, unless they are
empty or bad.  The first two items of the array should either be a p tag
or h2 tag.  This is a result of business process and cms limitations.